🔨 Refactor: SettingPage Jotai로 변경
