+++ src/pages/Setting/constants/settingTab.ts
@@ -1,7 +1,6 @@
 const SETTING_TAB = {
   LOGOUT: '로그아웃',
   INACTIVE: '커플 비활성화',
-  // CANCEL: '회원 탈퇴',
   REACTIVE: '커플 재활성화',
 } as const;
