--- src/pages/Setting/contexts/index.ts
@@ -1,1 +1,0 @@
-export { SettingContext, SettingProvider } from './SettingContext';
