--- src/pages/Setting/hooks/index.ts
@@ -1,1 +1,0 @@
-export { default as useSetting } from './useSetting';
