--- src/pages/Setting/hooks/useSetting.ts
@@ -1,12 +1,0 @@
-import { useContext } from 'react';
-import { SettingContext } from '../contexts';
-
-const useSetting = () => {
-  const settingContext = useContext(SettingContext);
-
-  if (!settingContext) throw new Error('Cannot find SettingProvider');
-
-  return settingContext;
-};
-
-export default useSetting;
