+++ src/pages/Setting/stores/settingAtom.ts
@@ -1,0 +1,4 @@
+import { atom } from 'jotai';
+import { SETTING_TAB } from '../constants';
+
+export const activeTabAtom = atom<keyof typeof SETTING_TAB>('LOGOUT');
